Add Texture Code from Class
Basic addition of texture, only works for the spheres used for the torches for now (might have to do with normalization, still looking into this). PPM file is required to load the desired texture, simply change the name of the ppm file to be loaded in the init function to change the texture. Also added glutWarpPointer command in main so mouse cursor is centered when the program is launched.